Update and rename .gitignore to NumberGuessingGame
This repository contains a Java application developed during a summer internship. The application is a number guessing game implemented using dialog boxes for user interaction. Users have five attempts to guess a randomly generated number. If the user fails to guess the correct number within the given attempts, the application reveals the correct number. Upon guessing the correct number, the user is congratulated. This project serves as a beginner-friendly example of Java GUI programming and basic game development.